requirements.txt
# LLM
anthropic>=0.40.0
ollama>=0.4.0

# Monitoring
langfuse>=2.0.0

# Storage & Encryption
cryptography>=42.0.0

# Market Data
yfinance>=0.2.0
requests>=2.32.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0

# UI
streamlit>=1.50.0
plotly>=5.0.0
pandas>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
nest_asyncio>=1.6.0
pydantic>=2.0.0
pyyaml>=6.0
